prediction-engine/requirements.txt
Flask==3.0.3
requests==2.32.3
grpcio==1.66.2
grpcio-tools==1.66.2
protobuf==5.28.2